+++ rust/CHANGELOG.md
@@ -16,6 +16,8 @@ This version requires leancrypto >= 1.8.0.
 
 * lcr_kyber_x25519 operation changed to provide non-KDF version usable for rustls
 
+* PBKDF2 interface
+
 ### Changed
 
 * API breakage: rename all getter functions to get_* (e.g. sk() -> get_sk())